subscriber/requirements.txt
flask
flask_restful
flask_cors
flask_socketio
requests
kafka-python
pymongo